Fixing a bug with multi remove query from ra-data-feathers restClient.
When using multi remove, ra-data-feathers restClient is sending this query:
{ _id: { '$in': [ '5e4cb5e3dd27050beb6eed2b' ] } }

Being one or more ids in the list, the query returns 0 found elements.

The problem seems to be that the query with ids enclosed by quotation marks
is not finding the documents that really exists, so we need to ensure each
id is valid, using ObjectID.

With this fix, the problem is solved and multi remove query format is working.